+++ tests/testthat/test-focus.R
@@ -7,34 +7,34 @@ d <- correlate(d)
 test_that("Column select works as dplyr::select", {
   expect_equal(
     colnames(focus(d, Sepal.Length, Sepal.Width, mirror = TRUE)),
-    c("rowname", "Sepal.Length", "Sepal.Width")
+    c("term", "Sepal.Length", "Sepal.Width")
   )
   expect_equal(
     colnames(focus(d, Sepal.Length:Petal.Length, mirror = TRUE)),
-    c("rowname", "Sepal.Length", "Sepal.Width", "Petal.Length")
+    c("term", "Sepal.Length", "Sepal.Width", "Petal.Length")
   )
   expect_equal(
     colnames(focus(d, dplyr::contains("Length"), mirror = TRUE)),
-    c("rowname", "Sepal.Length", "Petal.Length")
+    c("term", "Sepal.Length", "Petal.Length")
   )
   expect_equal(
     colnames(focus(d, -dplyr::contains("Length"), mirror = TRUE)),
-    c("rowname", "Sepal.Width", "Petal.Width")
+    c("term", "Sepal.Width", "Petal.Width")
   )
 })
 
 test_that("Selects/excludes in rows", {
   expect_equal(
-    focus(d, Sepal.Length, Sepal.Width, mirror = TRUE)$rowname,
+    focus(d, Sepal.Length, Sepal.Width, mirror = TRUE)$term,
     c("Sepal.Length", "Sepal.Width")
   )
   expect_equal(
-    focus(d, Sepal.Length, Sepal.Width, mirror = FALSE)$rowname,
+    focus(d, Sepal.Length, Sepal.Width, mirror = FALSE)$term,
     c("Petal.Length", "Petal.Width")
   )
   expect_equal(
     colnames(focus(d, Sepal.Length, Sepal.Width, mirror = FALSE)),
-    c("rowname", "Sepal.Length", "Sepal.Width")
+    c("term", "Sepal.Length", "Sepal.Width")
   )
   expect_equal(
     sum(is.na(as.matrix(focus(d, Sepal.Length, Sepal.Width, mirror = TRUE)))),